celery_queue/requirements.txt
celery==4.2.0
Pillow==5.4.1
redis==3.1.0
